stable but  have to draw after boolean
when cubies wants to be drawn without a boolean operation:
problem when Objects(), on cubie.cpp, gets called without doing the
boolean operation and the data pipeline to the cubies.
It is as if when first dawn, the group* objects gets corrupted.
error is here:
for (int j=0; j < ChCnt; j++){
//clone each object
sgCObject *temp = allParts[j]->Clone();
//put clone on *[] tomake new group
objcts[objctr] = temp;
objctr ++;
}
on the line:
sgCObject *temp = allParts[j]->Clone();
but all parts its giving a child count of 1.
HELP!